linkml-schema-manager/pyproject.toml
[project]
name = "linkml-schema-manager"
version = "0.1.0"
description = "FastAPI-based LinkML schema manager, validator, and toolkit"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "linkml>=1.8.0",
    "linkml-runtime>=1.8.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.0",
    "pydantic>=2.9.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "python-jose>=3.3.0",
    "difflib2>=0.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
